+++ matchms/similarity/__init__.py
@@ -6,7 +6,9 @@ Matchms provides a number of frequently used similarity scores to compare mass
 spectra. This includes
 
 * scores based on comparing peak positions and intensities
-  (:class:`~matchms.similarity.CosineGreedy` or :class:`~matchms.similarity.ModifiedCosine`)
+  (:class:`~matchms.similarity.CosineGreedy`,
+  :class:`~matchms.similarity.ModifiedCosineGreedy`,
+  :class:`~matchms.similarity.ModifiedCosineHungarian`)
 * simple scores that only assess precursor m/z or parent mass matches
   (:class:`~matchms.similarity.PrecursorMzMatch` or: :class:`~matchms.similarity.ParentMassMatch`)
 * scores assessing molecular similarity if structures (SMILES, InchiKey) are given as metadata
@@ -19,28 +21,34 @@ It is also easily possible to add own custom similarity measures or import exter
 (such as `Spec2Vec <https://github.com/iomega/spec2vec>`_).
 """
 from .BinnedEmbeddingSimilarity import BinnedEmbeddingSimilarity
+from .BlinkCosine import BlinkCosine
 from .CosineGreedy import CosineGreedy
 from .CosineHungarian import CosineHungarian
 from .FingerprintSimilarity import FingerprintSimilarity
+from .FlashSimilarity import FlashSimilarity
 from .IntersectMz import IntersectMz
 from .MetadataMatch import MetadataMatch
-from .ModifiedCosine import ModifiedCosine
+from .ModifiedCosineGreedy import ModifiedCosineGreedy
+from .ModifiedCosineHungarian import ModifiedCosineHungarian
 from .NeutralLossesCosine import NeutralLossesCosine
 from .ParentMassMatch import ParentMassMatch
 from .PrecursorMzMatch import PrecursorMzMatch
 
 
 __all__ = [
+    "BinnedEmbeddingSimilarity",
+    "BlinkCosine",
+    "ModifiedCosineGreedy",
     "CosineGreedy",
     "CosineHungarian",
     "FingerprintSimilarity",
+    "FlashSimilarity",
     "IntersectMz",
     "MetadataMatch",
-    "ModifiedCosine",
+    "ModifiedCosineHungarian",
     "NeutralLossesCosine",
     "ParentMassMatch",
     "PrecursorMzMatch",
-    "BinnedEmbeddingSimilarity"
 ]
 
 
@@ -54,8 +62,12 @@ def get_similarity_function_by_name(similarity_function_name: str):
         Name of the similarity function.
     """
     names = __all__
-    functions = [CosineGreedy, CosineHungarian, FingerprintSimilarity, IntersectMz, MetadataMatch, ModifiedCosine,
-                 NeutralLossesCosine, ParentMassMatch, PrecursorMzMatch, BinnedEmbeddingSimilarity]
+    functions = [BinnedEmbeddingSimilarity, BlinkCosine,
+                 ModifiedCosineGreedy,
+                 CosineGreedy, CosineHungarian,
+                 FingerprintSimilarity, FlashSimilarity,
+                 IntersectMz, MetadataMatch, ModifiedCosineHungarian,
+                 NeutralLossesCosine, ParentMassMatch, PrecursorMzMatch]
 
     assert similarity_function_name in names, f"Unknown similarity function: {similarity_function_name}"
     assert len(names) == len(functions), "Number of similarity functions and names do not match"